Move `ResourceInfo` to `references` for reuse in other modules
In the following commit, it will be used for resources preloading.
