+++ kmock/_internal/k8s.py
@@ -333,7 +333,7 @@ class KubernetesScaffold(apps.RawHandler):
     # The library itself DOES NOT use these fields, it uses the classes directly.
     # For when the fixture name overlaps the library name, so that it requires writing `import…as…`.
     ResourceKey = k8s_views.ResourceKey
-    ResourceInfo = k8s_views.ResourceInfo
+    ResourceInfo = references.ResourceInfo
     ResourceDict = k8s_views.ResourceDict
     ResourcesArray = k8s_views.ResourcesArray
